// dev/rAthenaImport/types/RAthenaMob.ts
/**
 * The data structure of the rAthena database table "mob_db"
 */
export type RAthenaMob = {
  ID: string;
  MVP1id: string;
  MVP1per: string;
  MVP2id: string;
  MVP2per: string;
  MVP3id: string;
  MVP3per: string;
  Drop1id: string;
  Drop1per: string;
  Drop2id: string;
  Drop2per: string;
  Drop3id: string;
  Drop3per: string;
  Drop4id: string;
  Drop4per: string;
  Drop5id: string;
  Drop5per: string;
  Drop6id: string;
  Drop6per: string;
  Drop7id: string;
  Drop7per: string;
  Drop8id: string;
  Drop8per: string;
  Drop9id: string;
  Drop9per: string;
  DropCardid: string;
  DropCardper: string;
};